+++ src/__mocks__/@react-native-community/async-storage.js
@@ -1,0 +1,3 @@
+import mockAsyncStorage from '@react-native-community/async-storage/jest/async-storage-mock'
+
+return mockAsyncStorage
