+++ src/__mocks__/@react-native-community/netinfo.js
@@ -1,0 +1,10 @@
+export default {
+  fetch: jest.fn(() => new Promise(resolve => resolve(true))),
+  getCurrentState: jest.fn(),
+  addEventListener: jest.fn(callback =>
+    callback({
+      isConnected: false
+    })
+  ),
+  removeListeners: jest.fn()
+}
