+++ src/__mocks__/react-native-device-info.js
@@ -1,0 +1,5 @@
+export default {
+  getVersion: () => '1.1.1',
+  getModel: jest.fn(),
+  getFreeDiskStorage: jest.fn()
+}
